+++ SerenityLib/examples/solarSystem/game/map/Freelist.h
@@ -9,7 +9,7 @@ template<typename T> class Freelist {
     private:
         std::vector<T>            m_Items;
         std::stack<unsigned int>  m_Freelist;
-        unsigned int              m_Size;
+        size_t                    m_Size;
     public:
         Freelist() {
             m_Size = 0;
@@ -20,21 +20,21 @@ template<typename T> class Freelist {
         ~Freelist() {
             clear(false);
         }
-        void initialize(const unsigned int _capacity) {
+        void initialize(const size_t _capacity) {
             m_Items.reserve(_capacity);
-            for (unsigned int i = 0; i < _capacity; ++i) {
+            for (size_t i = 0; i < _capacity; ++i) {
                 m_Items.push_back(nullptr);
             }
             for (int i = _capacity - 1; i >= 0; --i) {
                 m_Freelist.push(i);
             }
         }
-        const unsigned int& size() const {
+        const size_t& size() const {
             return m_Size;
         }
         void clear(const bool reinit = true) {
             auto cap = m_Items.capacity();
-            for (unsigned int i = 0; i < m_Items.size(); ++i) {
+            for (size_t i = 0; i < m_Items.size(); ++i) {
                 if (m_Items[i]) {
                     delete m_Items[i];
                     m_Items[i] = nullptr;
@@ -47,7 +47,7 @@ template<typename T> class Freelist {
                 initialize(cap);
         }
         const bool delete_data(T& data) {
-            for (unsigned int i = 0; i < m_Items.size(); ++i) {
+            for (size_t i = 0; i < m_Items.size(); ++i) {
                 if (m_Items[i] == data) {
                     delete m_Items[i];
                     m_Items[i] = nullptr;
@@ -58,7 +58,7 @@ template<typename T> class Freelist {
             }
             return false;
         }
-        const bool delete_data_index(const unsigned int index) {
+        const bool delete_data_index(const size_t index) {
             if (m_Items[index]) {
                 delete m_Items[index];
                 m_Items[index] = nullptr;
@@ -75,7 +75,7 @@ template<typename T> class Freelist {
             const auto available_index = m_Freelist.top();
             return available_index;
         }
-        const bool can_push_at_index(const int index) {
+        const bool can_push_at_index(const size_t index) {
             if (!m_Items[index])
                 return true;
             return false;
@@ -98,7 +98,7 @@ template<typename T> class Freelist {
             ++m_Size;
             return available_index;
         }
-        const bool insert(T& data, const int index) {
+        const bool insert(T& data, const size_t index) {
             if (!m_Items[index]){
                 m_Items[index] = std::move(data);
                 ++m_Size;
@@ -106,7 +106,7 @@ template<typename T> class Freelist {
             }
             return false;
         }
-        T& operator[](const unsigned int index) {
+        T& operator[](const size_t index) {
             return m_Items[index];
         }
         std::vector<T>& data() {
